backend/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# IA et traitement de texte
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
python-docx==1.1.0
jinja2==3.1.2
weasyprint==60.2

# Utilitaires
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Développement
black==23.11.0
isort==5.12.0
flake8==6.1.0 